// src/app/models/category.ts
export class Category {
    [x: string]: any;
    id?: string;
    category?: string;
}

export interface Position {
    id?: string;
    position?: string;
}

export interface Setup {
    id?: string;
    setup?: string;
}

export interface Rate {
    id?: string;
    rate?: string;
}

export interface Skill {
    id?: string;
    itemName?: string;
}

export interface Degree {
    id?: string;
    degree?: string;
}

export interface Course {
    id?: string;
    course?: string;
}

export interface University {
    id?: string;
    university?: string;
}